Fix: util에 있던 api 호출 전부 액티비티로 이동
