+++ app/src/main/java/com/example/spoot_taxi_front/activities/MainActivity.java
@@ -20,6 +20,7 @@ import android.os.Build;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.MenuItem;
+import android.widget.Toast;
 
 
 import com.example.spoot_taxi_front.databinding.ActivityMainBinding;
@@ -30,15 +31,32 @@ import com.example.spoot_taxi_front.fragments.RallyFragment;
 import com.example.spoot_taxi_front.fragments.SettingsFragment;
 
 
+import com.example.spoot_taxi_front.models.ChatRoom;
+import com.example.spoot_taxi_front.models.User;
+import com.example.spoot_taxi_front.network.api.ChatApi;
+import com.example.spoot_taxi_front.network.dto.UserDto;
+import com.example.spoot_taxi_front.network.dto.UserJoinedChatRoomDto;
+import com.example.spoot_taxi_front.network.dto.responses.UserJoinedChatRoomResponse;
+import com.example.spoot_taxi_front.network.retrofit.ApiManager;
 import com.example.spoot_taxi_front.network.socket.WebSocketViewModel;
 import com.example.spoot_taxi_front.utils.LocalChatRoomManager;
 import com.example.spoot_taxi_front.utils.ChatRoomDataChange;
+import com.example.spoot_taxi_front.utils.SessionManager;
 import com.google.android.material.badge.BadgeDrawable;
 import com.kakao.util.maps.helper.Utility;
 
 import org.greenrobot.eventbus.EventBus;
 import org.greenrobot.eventbus.Subscribe;
 import org.greenrobot.eventbus.ThreadMode;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
+
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -69,7 +87,7 @@ public class MainActivity extends AppCompatActivity {
         binding = ActivityMainBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
 
-        localChatRoomManager.loadChatRoomsFromServer();
+        loadChatRoomsFromServer();
 
         // 채팅 아이템의 Badge 설정
         MenuItem chatMenuItem = binding.navigationView.getMenu().findItem(R.id.chatFragment);
@@ -140,7 +158,7 @@ public class MainActivity extends AppCompatActivity {
         if(!webSocketViewModel.isConnected()){
             Log.d("resumeDisConnect","비연결되어있어요");
             webSocketViewModel.reconnect();
-            LocalChatRoomManager.getInstance().loadChatRoomsFromServer();
+            loadChatRoomsFromServer();
         }
     }
 
@@ -270,5 +288,70 @@ public class MainActivity extends AppCompatActivity {
                             }
                         }
                     });
+
+
+    public void loadChatRoomsFromServer() {
+
+        ChatApi chatApi = ApiManager.getInstance().createChatApi(SessionManager.getInstance().getJwtToken());
+        Call<UserJoinedChatRoomResponse> call = chatApi.getUserChatRooms(SessionManager.getInstance().getCurrentUser().getEmail());
+        call.enqueue(new Callback<UserJoinedChatRoomResponse>() {
+            @Override
+            public void onResponse(Call<UserJoinedChatRoomResponse> call, Response<UserJoinedChatRoomResponse> response) {
+                localChatRoomManager.setChatRooms(extractChatRoomListFromResponse(response.code(), response.body()));
+                EventBus.getDefault().post(new ChatRoomDataChange());
+            }
+
+            @Override
+            public void onFailure(Call<UserJoinedChatRoomResponse> call, Throwable t) {
+                Log.e("API Failure", "API 호출에 실패하였습니다.", t);
+            }
+        });
+
+    }
+
+    private List<ChatRoom> extractChatRoomListFromResponse(int statusCode, UserJoinedChatRoomResponse responseBody) {
+        switch (statusCode) {
+            case 200:
+                List<UserJoinedChatRoomDto> userJoinedChatRoomDtoList = responseBody.getUserJoinedChatRoomDtoList();
+                return parseDtoToChatRooms(userJoinedChatRoomDtoList);
+            case 403:
+                Toast.makeText(getApplicationContext(), "서비스 이용을 위해 재로그인 해주세요", Toast.LENGTH_SHORT).show();
+                Intent reAuthneticateIntent = new Intent(getApplicationContext(), LoginActivity.class);
+                startActivity(reAuthneticateIntent);
+            default:
+                break;
+        }
+        return null;
+    }
+
+    private List<ChatRoom> parseDtoToChatRooms(List<UserJoinedChatRoomDto> userJoinedChatRoomDtoList) {
+        List<ChatRoom> chatRoomApiResponseList = new ArrayList<>();
+        for (UserJoinedChatRoomDto userJoinedChatRoomDto : userJoinedChatRoomDtoList) {
+            Long chatRoomId = userJoinedChatRoomDto.getChatRoomId();
+            String chatRoomName = userJoinedChatRoomDto.getChatRoomName();
+
+            List<User> userList = new ArrayList<>();
+            List<UserDto> participants = userJoinedChatRoomDto.getParticipants();
+            for (UserDto participant : participants) {
+                User user = new User(participant.getEmail(), participant.getPassword(), participant.getName(), participant.getImgUrl(), participant.getGender());
+                userList.add(user);
+            }
+
+            Optional<String> optionalLastMessage = Optional.ofNullable(userJoinedChatRoomDto.getLastMessage());
+            String lastMessage = optionalLastMessage.orElse("");
+
+            //LocalDateTime lastSentTime = userJoinedChatRoomDto.getLastSentTime();
+
+            Optional<LocalDateTime> optionalLastSentTime = Optional.ofNullable(userJoinedChatRoomDto.getLastSentTime());
+            String lastSentTimeString = optionalLastSentTime.map(LocalDateTime::toString).orElse("");
+            Integer nonReadMessageCount = userJoinedChatRoomDto.getNonReadMessageCount();
+            ChatRoom chatRoom = new ChatRoom(chatRoomId,chatRoomName,userList,lastMessage,lastSentTimeString,nonReadMessageCount);
+            Log.d("채팅방 목록",chatRoom.toString());
+            chatRoomApiResponseList.add(chatRoom);
+            // 특정 채널 구독
+            webSocketViewModel.subscribeToChannel(chatRoomId);
+        }
+        return chatRoomApiResponseList;
+    }
 
 }
