+++ app/src/main/java/com/example/spoot_taxi_front/network/socket/WebSocketManager.java
@@ -31,7 +31,9 @@ public class WebSocketManager {
 
     // 구독한 주소와 해당 구독의 Disposable을 매핑하는 Map을 사용
     Map<String, Disposable> subscriptionMap = new HashMap<>();
-    private static final String WEBSOCKET_URL = "ws://192.168.219.106:8090/ws/websocket";
+//    private static final String WEBSOCKET_URL = "ws://192.168.219.106:8090/ws/websocket";
+    private static final String WEBSOCKET_URL = "ws://192.168.219.110:8090/ws/websocket";
+
     private WebSocketManager() {
         // Private constructor to prevent external instantiation
     }
