+++ app/src/main/java/com/example/spoot_taxi_front/utils/LocalChatRoomManager.java
@@ -107,63 +107,5 @@ public class LocalChatRoomManager {
         }
         return res;
     }
-    public void loadChatRoomsFromServer() {
 
-        Call<UserJoinedChatRoomResponse> call = chatApi.getUserChatRooms(SessionManager.getInstance().getCurrentUser().getEmail());
-
-        call.enqueue(new Callback<UserJoinedChatRoomResponse>() {
-            @Override
-            public void onResponse(Call<UserJoinedChatRoomResponse> call, Response<UserJoinedChatRoomResponse> response) {
-                chatRooms = extractChatRoomListFromResponse(response.code(), response.body());
-                EventBus.getDefault().post(new ChatRoomDataChange());
-            }
-
-            @Override
-            public void onFailure(Call<UserJoinedChatRoomResponse> call, Throwable t) {
-                Log.e("API Failure", "API 호출에 실패하였습니다.", t);
-            }
-        });
-
-    }
-
-    private List<ChatRoom> extractChatRoomListFromResponse(int statusCode, UserJoinedChatRoomResponse responseBody) {
-        switch (statusCode) {
-            case 200:
-                List<UserJoinedChatRoomDto> userJoinedChatRoomDtoList = responseBody.getUserJoinedChatRoomDtoList();
-                return parseDtoToChatRooms(userJoinedChatRoomDtoList);
-            default:
-                break;
-        }
-        return null;
-    }
-
-    private List<ChatRoom> parseDtoToChatRooms(List<UserJoinedChatRoomDto> userJoinedChatRoomDtoList) {
-        List<ChatRoom> chatRoomApiResponseList = new ArrayList<>();
-        for (UserJoinedChatRoomDto userJoinedChatRoomDto : userJoinedChatRoomDtoList) {
-            Long chatRoomId = userJoinedChatRoomDto.getChatRoomId();
-            String chatRoomName = userJoinedChatRoomDto.getChatRoomName();
-
-            List<User> userList = new ArrayList<>();
-            List<UserDto> participants = userJoinedChatRoomDto.getParticipants();
-            for (UserDto participant : participants) {
-                User user = new User(participant.getEmail(), participant.getPassword(), participant.getName(), participant.getImgUrl(), participant.getGender());
-                userList.add(user);
-            }
-
-            Optional<String> optionalLastMessage = Optional.ofNullable(userJoinedChatRoomDto.getLastMessage());
-            String lastMessage = optionalLastMessage.orElse("");
-
-            //LocalDateTime lastSentTime = userJoinedChatRoomDto.getLastSentTime();
-
-            Optional<LocalDateTime> optionalLastSentTime = Optional.ofNullable(userJoinedChatRoomDto.getLastSentTime());
-            String lastSentTimeString = optionalLastSentTime.map(LocalDateTime::toString).orElse("");
-            Integer nonReadMessageCount = userJoinedChatRoomDto.getNonReadMessageCount();
-            ChatRoom chatRoom = new ChatRoom(chatRoomId,chatRoomName,userList,lastMessage,lastSentTimeString,nonReadMessageCount);
-            Log.d("채팅방 목록",chatRoom.toString());
-            chatRoomApiResponseList.add(chatRoom);
-            // 특정 채널 구독
-            webSocketViewModel.subscribeToChannel(chatRoomId);
-        }
-        return chatRoomApiResponseList;
-    }
 }
